Unify AirPlay streaming on a single cliairplay binary
Replace the separate cliraop (RAOP) and cliap2 (owntone-based AirPlay 2)
binaries with a single cliairplay binary that speaks both protocols,
selected via a --protocol flag.

- Drop the protocols/ package (RaopStream/AirPlay2Stream/_protocol) in
  favour of one unified AirPlayStream in stream.py.
- Remove the encryption/ALAC-encode config toggles (the binary always
  ALAC-encodes) and the AirPlay 2 "no sync" warning, since AP2 sync is
  now handled by the new binary.
- get_cli_binary() no longer takes a protocol argument.
