+++ music_assistant/providers/airplay/constants.py
@@ -21,14 +21,11 @@ class StreamingProtocol(IntEnum):
     AIRPLAY2 = 2  # AirPlay 2
 
 
-CONF_ENCRYPTION: Final[str] = "encryption"
-CONF_ALAC_ENCODE: Final[str] = "alac_encode"
 CONF_VOLUME_START: Final[str] = "volume_start"
 CONF_SESSION_ESTABLISHMENT_LATENCY: Final[str] = "session_establishment_latency"
 CONF_PASSWORD: Final[str] = "password"
 CONF_AP2PASSWORD: Final[str] = "ap2password"
 CONF_IGNORE_VOLUME: Final[str] = "ignore_volume"
-CONF_CREDENTIALS: Final[str] = "credentials"
 CONF_AIRPLAY_PROTOCOL: Final[str] = "airplay_protocol"
 CONF_STORED_VOLUME: Final[str] = "stored_volume"
 
@@ -38,7 +35,7 @@ DACP_DISCOVERY_TYPE: Final[str] = "_dacp._tcp.local."
 
 # Time allowance for MA processing and OS spawning of the binary. Helps reduce initial audio loss.
 AIRPLAY_DEFAULT_SESSION_DELAY_MS: Final[int] = 900
-# Read ahead buffer for cliraop. Default output buffer duration.
+# Read ahead buffer. Default output buffer duration.
 AIRPLAY_OUTPUT_BUFFER_DEFAULT_DURATION_MS: Final[int] = 1500
 # RAOP buffer settings
 RAOP_OUTPUT_BUFFER_MIN_DURATION_MS: Final[int] = 250
@@ -51,7 +48,6 @@ AIRPLAY_SESSION_ESTABLISHMENT_LATENCY_MIN_MS: Final[int] = (
 AIRPLAY_SESSION_ESTABLISHMENT_LATENCY_MAX_MS: Final[int] = (
     4000  # Maximum session establishment latency permitted
 )
-AIRPLAY2_MIN_LOG_LEVEL: Final[int] = 3  # Min loglevel to ensure stderr output contains what we need
 RAOP_CONNECT_TIME_MS: Final[int] = 1500  # Time in ms to allow RAOP device to connect
 
 # Per-protocol credential storage keys
@@ -60,9 +56,6 @@ CONF_AIRPLAY_CREDENTIALS: Final[str] = "airplay_credentials"
 
 # Some RAOP models require a higher than default 1000ms buffer to prevent stuttering
 CONF_RAOP_LATENCY: Final[str] = "airplay_latency"
-
-# Legacy credential key (for migration)
-CONF_AP_CREDENTIALS: Final[str] = "ap_credentials"
 
 # Pairing action keys
 CONF_ACTION_START_PAIRING: Final[str] = "start_pairing"
