--- music_assistant/providers/airplay/protocols/__init__.py
@@ -1,1 +1,0 @@
-"""AirPlay (streaming) Protocols."""
